+++ tgbotapi.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/ClassCastsNew.kt
@@ -161,7 +161,6 @@ import dev.inmo.tgbotapi.types.chat.UnknownExtendedChat
 import dev.inmo.tgbotapi.types.chat.User
 import dev.inmo.tgbotapi.types.chat.UsernameChat
 import dev.inmo.tgbotapi.types.chat.member.AdministratorChatMember
-import dev.inmo.tgbotapi.types.chat.member.BannedChatMember
 import dev.inmo.tgbotapi.types.chat.member.ChatMember
 import dev.inmo.tgbotapi.types.chat.member.ChatMemberUpdated
 import dev.inmo.tgbotapi.types.chat.member.KickedChatMember
@@ -690,15 +689,6 @@ public inline fun OptionallyWithUser.administratorChatMemberOrThrow(): Administr
 public inline fun <T>
     OptionallyWithUser.ifAdministratorChatMember(block: (AdministratorChatMember) -> T): T? =
     administratorChatMemberOrNull() ?.let(block)
-
-public inline fun OptionallyWithUser.bannedChatMemberOrNull(): BannedChatMember? = this as?
-    dev.inmo.tgbotapi.types.chat.member.BannedChatMember
-
-public inline fun OptionallyWithUser.bannedChatMemberOrThrow(): BannedChatMember = this as
-    dev.inmo.tgbotapi.types.chat.member.BannedChatMember
-
-public inline fun <T> OptionallyWithUser.ifBannedChatMember(block: (BannedChatMember) -> T): T? =
-    bannedChatMemberOrNull() ?.let(block)
 
 public inline fun OptionallyWithUser.chatMemberOrNull(): ChatMember? = this as?
     dev.inmo.tgbotapi.types.chat.member.ChatMember