+++ CHANGELOG.md
@@ -2,12 +2,34 @@
 
 ## 18.0.0
 
+**THIS UPDATE CONTAINS BREAKING CHANGES**
+
+---
+
+**Migration (step-by-step)**
+
+Replace in a whole project:
+
+1. `RestrictedChatMember` -> `RestrictedMemberChatMember`
+2. `BannedChatMember` -> `RestrictedChatMember`
+
+---
+
+**Changes:**
+
 * `Version`:
   * `Kotlin`: `2.0.10` -> `2.0.20`
   * `Serialization`: `1.7.1` -> `1.7.2`
   * `MicroUtils`: `0.22.0` -> `0.22.1`
   * `KSLog`: `1.3.5` -> `1.3.6`
   * `Ktor`: `2.3.11` -> `2.3.12`
+* `Core`:
+  * `RestrictedChatMember` has been renamed to `RestrictedMemberChatMember`
+  * `BannedChatMember` has been renamed to `RestrictedChatMember`
+  * `KickedChatMember` now implements `RestrictedChatMember` (due to rename)
+* `Utils`:
+  * Add a lot of extensions like `ChatMember.isMember`
+  * Add a lot of extensions like `ChatMemberUpdated.joinedChat`
 
 ## 17.0.0